Move params merging tests to own file
